LRAC-12616 Create '__tests__' folder to People and move files to it
