+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/components/people/__tests__/AccountGroupsModal.spec.tsx
@@ -16,57 +16,28 @@ import fetch from 'jest-fetch-mock';
 
 import '@testing-library/jest-dom/extend-expect';
 import {useModal} from '@clayui/modal';
-import {
-	act,
-	cleanup,
-	fireEvent,
-	render,
-	screen,
-	waitFor,
-	within,
-} from '@testing-library/react';
+import {act, cleanup, render, screen, waitFor} from '@testing-library/react';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import {TEmptyState} from '../table/StateRenderer';
-import {TTableRequestParams} from '../table/types';
-import Modal from './Modal';
-import {EPeople} from './People';
+import {TEmptyState} from '../../table/StateRenderer';
+import {TTableRequestParams} from '../../table/types';
+import Modal from '../Modal';
+import {EPeople} from '../People';
 
 const responseWithData = {
 	actions: {},
 	facets: [],
 	items: [
 		{
-			id: 44536,
-			name: 'organizations test',
+			id: 10101,
+			name: 'account groups test',
 			selected: false,
 		},
 		{
-			id: 44542,
-			name: 'organizations test 2',
+			id: 21210,
+			name: 'account groups test 2',
 			selected: false,
-		},
-	],
-	lastPage: 1,
-	page: 1,
-	pageSize: 20,
-	totalCount: 2,
-};
-
-const responseWithDataAllSelected = {
-	actions: {},
-	facets: [],
-	items: [
-		{
-			id: 44536,
-			name: 'organizations test',
-			selected: true,
-		},
-		{
-			id: 44542,
-			name: 'organizations test 2',
-			selected: true,
 		},
 	],
 	lastPage: 1,
@@ -109,11 +80,11 @@ const ComponentWithData: React.FC<IComponentWithDataProps> = ({requestFn}) => {
 				{
 					expanded: true,
 					id: 'name',
-					label: Liferay.Language.get('organizations'),
+					label: Liferay.Language.get('account-group'),
 				},
 			]}
 			emptyState={emptyState}
-			name={EPeople.OrganizationIds}
+			name={EPeople.UserGroupIds}
 			observer={observer}
 			onCloseModal={() => {}}
 			requestFn={requestFn}
@@ -138,14 +109,14 @@ const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
 		contentRenderer: () => <></>,
 		description: 'Empty State Description',
 		noResultsTitle: 'Empty State No Results Title',
-		title: 'There are no organizations',
+		title: 'There are no account groups',
 	};
 
 	return (
 		<Modal
 			columns={[]}
 			emptyState={emptyState}
-			name={EPeople.OrganizationIds}
+			name={EPeople.UserGroupIds}
 			observer={observer}
 			onCloseModal={() => {}}
 			requestFn={requestFn}
@@ -156,14 +127,13 @@ const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
 				syncedOrganizationIds: [''],
 				syncedUserGroupIds: [''],
 			}}
-			title="Add Organizations"
+			title="Add Account Groups"
 		/>
 	);
 };
 
-describe('Organizations Modal', () => {
+describe('Account Groups Modal', () => {
 	beforeAll(() => {
-
 		// @ts-ignore
 
 		ReactDOM.createPortal = jest.fn((element) => {
@@ -182,7 +152,7 @@ describe('Organizations Modal', () => {
 		cleanup();
 	});
 
-	it('renders Organizations Modal with data without crashing it', async () => {
+	it('renders Account Groups modal without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithData));
 
 		await act(async () => {
@@ -192,16 +162,16 @@ describe('Organizations Modal', () => {
 
 			jest.runAllTimers();
 
-			await waitFor(() => screen.getByText('organizations test'));
+			await waitFor(() => screen.getByText('account groups test'));
 
-			await waitFor(() => screen.getByText('organizations test 2'));
+			await waitFor(() => screen.getByText('account groups test 2'));
 		});
 
 		const modalContent = document.querySelector('.modal-content');
 
-		const tableColumnText = screen.getByText('organizations test');
+		const tableColumnText = screen.getByText('account groups test');
 
-		const tableColumnText2 = screen.getByText('organizations test 2');
+		const tableColumnText2 = screen.getByText('account groups test 2');
 
 		expect(modalContent).toBeInTheDocument();
 
@@ -210,7 +180,7 @@ describe('Organizations Modal', () => {
 		expect(tableColumnText2).toBeInTheDocument();
 	});
 
-	it('renders Organizations Modal with Empty State without crashing it', async () => {
+	it('renders User Groups Modal with Empty State without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithEmptyState));
 
 		await act(async () => {
@@ -222,54 +192,18 @@ describe('Organizations Modal', () => {
 			jest.runAllTimers();
 		});
 
-		const organizationsTitle = screen.getByText('Add Organizations');
+		const accountGroupsTitle = screen.getByText('Add Account Groups');
 
-		const emptyStateTitle = screen.getByText('There are no organizations');
+		const emptyStateTitle = screen.getByText('There are no account groups');
 
 		const emptyStateDescription = screen.getByText(
 			'Empty State Description'
 		);
 
-		expect(organizationsTitle).toBeInTheDocument();
+		expect(accountGroupsTitle).toBeInTheDocument();
 
 		expect(emptyStateTitle).toBeInTheDocument();
 
 		expect(emptyStateDescription).toBeInTheDocument();
-	});
-
-	//  TODO: Refactor the test below to be able to fetch the mocked data correctly. Only passes when it runs alone.
-
-	it.skip('renders Organizations Modal, click on checkbox to select all items', async () => {
-		fetch.mockResponse(JSON.stringify(responseWithData));
-
-		await act(async () => {
-			render(
-				<ComponentWithData requestFn={async () => responseWithData} />
-			);
-
-			jest.runAllTimers();
-		});
-
-		const navigation = screen.getByRole('navigation');
-
-		const selectAllCheckboxes = within(navigation).getByRole('checkbox');
-
-		await act(async () => {
-			fireEvent.click(selectAllCheckboxes);
-		});
-
-		fetch.mockResponse(JSON.stringify(responseWithDataAllSelected));
-
-		const tableRow = document.querySelector(
-			'tr[data-testid="organizations test"]'
-		);
-
-		const tableRow2 = document.querySelector(
-			'tr[data-testid="organizations test 2"]'
-		);
-
-		expect(tableRow).toHaveClass('table-active');
-
-		expect(tableRow2).toHaveClass('table-active');
 	});
 });
