+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/components/people/__tests__/Modal.spec.tsx
@@ -20,30 +20,29 @@ import {act, cleanup, render, screen, waitFor} from '@testing-library/react';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import {TEmptyState} from '../table/StateRenderer';
-import {TTableRequestParams} from '../table/types';
-import Modal from './Modal';
-import {EPeople} from './People';
+import {TEmptyState} from '../../table/StateRenderer';
+import Modal from '../Modal';
+import {EPeople} from '../People';
 
 const responseWithData = {
 	actions: {},
 	facets: [],
 	items: [
 		{
-			id: 10101,
-			name: 'account groups test',
+			id: 44275,
+			name: 'test',
 			selected: false,
 		},
 		{
-			id: 21210,
-			name: 'account groups test 2',
+			id: 44279,
+			name: 'test2',
 			selected: false,
 		},
 	],
 	lastPage: 1,
 	page: 1,
 	pageSize: 20,
-	totalCount: 2,
+	totalCount: 7,
 };
 
 const responseWithEmptyState = {
@@ -56,15 +55,7 @@ const responseWithEmptyState = {
 	totalCount: 0,
 };
 
-interface IComponentWithDataProps {
-	requestFn: (params: TTableRequestParams) => Promise<any>;
-}
-
-interface IComponentWithEmptyStateProps {
-	requestFn: (params: TTableRequestParams) => Promise<any>;
-}
-
-const ComponentWithData: React.FC<IComponentWithDataProps> = ({requestFn}) => {
+const ComponentWithData = () => {
 	const {observer} = useModal({onClose: () => {}});
 
 	const emptyState: TEmptyState = {
@@ -80,36 +71,34 @@ const ComponentWithData: React.FC<IComponentWithDataProps> = ({requestFn}) => {
 				{
 					expanded: true,
 					id: 'name',
-					label: Liferay.Language.get('account-group'),
+					label: Liferay.Language.get('user-groups'),
 				},
 			]}
 			emptyState={emptyState}
 			name={EPeople.UserGroupIds}
 			observer={observer}
 			onCloseModal={() => {}}
-			requestFn={requestFn}
+			requestFn={async () => responseWithData}
 			syncAllAccounts
 			syncAllContacts
 			syncedIds={{
-				syncedAccountGroupIds: [''],
-				syncedOrganizationIds: [''],
-				syncedUserGroupIds: [''],
+				syncedAccountGroupIds: ['aaa'],
+				syncedOrganizationIds: ['bbb'],
+				syncedUserGroupIds: ['ccc'],
 			}}
 			title="Assign Modal Title"
 		/>
 	);
 };
 
-const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
-	requestFn,
-}) => {
+const ComponentWithEmptyState = () => {
 	const {observer} = useModal({onClose: () => {}});
 
 	const emptyState: TEmptyState = {
 		contentRenderer: () => <></>,
 		description: 'Empty State Description',
 		noResultsTitle: 'Empty State No Results Title',
-		title: 'There are no account groups',
+		title: 'Empty State Title',
 	};
 
 	return (
@@ -119,27 +108,27 @@ const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
 			name={EPeople.UserGroupIds}
 			observer={observer}
 			onCloseModal={() => {}}
-			requestFn={requestFn}
+			requestFn={async () => responseWithEmptyState}
 			syncAllAccounts
 			syncAllContacts
 			syncedIds={{
-				syncedAccountGroupIds: [''],
-				syncedOrganizationIds: [''],
-				syncedUserGroupIds: [''],
+				syncedAccountGroupIds: ['aaa'],
+				syncedOrganizationIds: ['bbb'],
+				syncedUserGroupIds: ['ccc'],
 			}}
-			title="Add Account Groups"
+			title="Assign Modal Title"
 		/>
 	);
 };
 
-describe('Account Groups Modal', () => {
+describe('People Modal', () => {
 	beforeAll(() => {
-
 		// @ts-ignore
 
 		ReactDOM.createPortal = jest.fn((element) => {
 			return element;
 		});
+
 		jest.useFakeTimers();
 	});
 
@@ -153,26 +142,24 @@ describe('Account Groups Modal', () => {
 		cleanup();
 	});
 
-	it('renders Account Groups modal without crashing it', async () => {
+	it('renders component with data without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithData));
 
 		await act(async () => {
-			render(
-				<ComponentWithData requestFn={async () => responseWithData} />
-			);
+			render(<ComponentWithData />);
 
 			jest.runAllTimers();
 
-			await waitFor(() => screen.getByText('account groups test'));
+			await waitFor(() => screen.getByText('test'));
 
-			await waitFor(() => screen.getByText('account groups test 2'));
+			await waitFor(() => screen.getByText('test2'));
 		});
 
 		const modalContent = document.querySelector('.modal-content');
 
-		const tableColumnText = screen.getByText('account groups test');
+		const tableColumnText = screen.getByText('test');
 
-		const tableColumnText2 = screen.getByText('account groups test 2');
+		const tableColumnText2 = screen.getByText('test2');
 
 		expect(modalContent).toBeInTheDocument();
 
@@ -181,27 +168,23 @@ describe('Account Groups Modal', () => {
 		expect(tableColumnText2).toBeInTheDocument();
 	});
 
-	it('renders User Groups Modal with Empty State without crashing it', async () => {
+	it('renders component with Empty State without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithEmptyState));
 
 		await act(async () => {
-			render(
-				<ComponentWithEmptyState
-					requestFn={async () => responseWithEmptyState}
-				/>
-			);
+			render(<ComponentWithEmptyState />);
 			jest.runAllTimers();
 		});
 
-		const accountGroupsTitle = screen.getByText('Add Account Groups');
+		const assignModalTitle = screen.getByText('Assign Modal Title');
 
-		const emptyStateTitle = screen.getByText('There are no account groups');
+		const emptyStateTitle = screen.getByText('Empty State Title');
 
 		const emptyStateDescription = screen.getByText(
 			'Empty State Description'
 		);
 
-		expect(accountGroupsTitle).toBeInTheDocument();
+		expect(assignModalTitle).toBeInTheDocument();
 
 		expect(emptyStateTitle).toBeInTheDocument();
