+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/components/people/__tests__/OrganizationsModal.spec.tsx
@@ -16,28 +16,57 @@ import fetch from 'jest-fetch-mock';
 
 import '@testing-library/jest-dom/extend-expect';
 import {useModal} from '@clayui/modal';
-import {act, cleanup, render, screen, waitFor} from '@testing-library/react';
+import {
+	act,
+	cleanup,
+	fireEvent,
+	render,
+	screen,
+	waitFor,
+	within,
+} from '@testing-library/react';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import {TEmptyState} from '../table/StateRenderer';
-import {TTableRequestParams} from '../table/types';
-import Modal from './Modal';
-import {EPeople} from './People';
+import {TEmptyState} from '../../table/StateRenderer';
+import {TTableRequestParams} from '../../table/types';
+import Modal from '../Modal';
+import {EPeople} from '../People';
 
 const responseWithData = {
 	actions: {},
 	facets: [],
 	items: [
 		{
-			id: 11100,
-			name: 'user groups test',
+			id: 44536,
+			name: 'organizations test',
 			selected: false,
 		},
 		{
-			id: 12120,
-			name: 'user groups test 2',
+			id: 44542,
+			name: 'organizations test 2',
 			selected: false,
+		},
+	],
+	lastPage: 1,
+	page: 1,
+	pageSize: 20,
+	totalCount: 2,
+};
+
+const responseWithDataAllSelected = {
+	actions: {},
+	facets: [],
+	items: [
+		{
+			id: 44536,
+			name: 'organizations test',
+			selected: true,
+		},
+		{
+			id: 44542,
+			name: 'organizations test 2',
+			selected: true,
 		},
 	],
 	lastPage: 1,
@@ -80,11 +109,11 @@ const ComponentWithData: React.FC<IComponentWithDataProps> = ({requestFn}) => {
 				{
 					expanded: true,
 					id: 'name',
-					label: Liferay.Language.get('user-groups'),
+					label: Liferay.Language.get('organizations'),
 				},
 			]}
 			emptyState={emptyState}
-			name={EPeople.UserGroupIds}
+			name={EPeople.OrganizationIds}
 			observer={observer}
 			onCloseModal={() => {}}
 			requestFn={requestFn}
@@ -109,14 +138,14 @@ const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
 		contentRenderer: () => <></>,
 		description: 'Empty State Description',
 		noResultsTitle: 'Empty State No Results Title',
-		title: 'There are no user groups',
+		title: 'There are no organizations',
 	};
 
 	return (
 		<Modal
 			columns={[]}
 			emptyState={emptyState}
-			name={EPeople.UserGroupIds}
+			name={EPeople.OrganizationIds}
 			observer={observer}
 			onCloseModal={() => {}}
 			requestFn={requestFn}
@@ -127,14 +156,13 @@ const ComponentWithEmptyState: React.FC<IComponentWithEmptyStateProps> = ({
 				syncedOrganizationIds: [''],
 				syncedUserGroupIds: [''],
 			}}
-			title="Add User Groups"
+			title="Add Organizations"
 		/>
 	);
 };
 
-describe('User Groups Modal', () => {
+describe('Organizations Modal', () => {
 	beforeAll(() => {
-
 		// @ts-ignore
 
 		ReactDOM.createPortal = jest.fn((element) => {
@@ -153,7 +181,7 @@ describe('User Groups Modal', () => {
 		cleanup();
 	});
 
-	it('renders User Groups modal without crashing it', async () => {
+	it('renders Organizations Modal with data without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithData));
 
 		await act(async () => {
@@ -163,16 +191,16 @@ describe('User Groups Modal', () => {
 
 			jest.runAllTimers();
 
-			await waitFor(() => screen.getByText('user groups test'));
+			await waitFor(() => screen.getByText('organizations test'));
 
-			await waitFor(() => screen.getByText('user groups test 2'));
+			await waitFor(() => screen.getByText('organizations test 2'));
 		});
 
 		const modalContent = document.querySelector('.modal-content');
 
-		const tableColumnText = screen.getByText('user groups test');
+		const tableColumnText = screen.getByText('organizations test');
 
-		const tableColumnText2 = screen.getByText('user groups test 2');
+		const tableColumnText2 = screen.getByText('organizations test 2');
 
 		expect(modalContent).toBeInTheDocument();
 
@@ -181,7 +209,7 @@ describe('User Groups Modal', () => {
 		expect(tableColumnText2).toBeInTheDocument();
 	});
 
-	it('renders User Groups Modal with Empty State without crashing it', async () => {
+	it('renders Organizations Modal with Empty State without crashing it', async () => {
 		fetch.mockResponse(JSON.stringify(responseWithEmptyState));
 
 		await act(async () => {
@@ -193,18 +221,54 @@ describe('User Groups Modal', () => {
 			jest.runAllTimers();
 		});
 
-		const userGroupsTitle = screen.getByText('Add User Groups');
+		const organizationsTitle = screen.getByText('Add Organizations');
 
-		const emptyStateTitle = screen.getByText('There are no user groups');
+		const emptyStateTitle = screen.getByText('There are no organizations');
 
 		const emptyStateDescription = screen.getByText(
 			'Empty State Description'
 		);
 
-		expect(userGroupsTitle).toBeInTheDocument();
+		expect(organizationsTitle).toBeInTheDocument();
 
 		expect(emptyStateTitle).toBeInTheDocument();
 
 		expect(emptyStateDescription).toBeInTheDocument();
+	});
+
+	//  TODO: Refactor the test below to be able to fetch the mocked data correctly. Only passes when it runs alone.
+
+	it.skip('renders Organizations Modal, click on checkbox to select all items', async () => {
+		fetch.mockResponse(JSON.stringify(responseWithData));
+
+		await act(async () => {
+			render(
+				<ComponentWithData requestFn={async () => responseWithData} />
+			);
+
+			jest.runAllTimers();
+		});
+
+		const navigation = screen.getByRole('navigation');
+
+		const selectAllCheckboxes = within(navigation).getByRole('checkbox');
+
+		await act(async () => {
+			fireEvent.click(selectAllCheckboxes);
+		});
+
+		fetch.mockResponse(JSON.stringify(responseWithDataAllSelected));
+
+		const tableRow = document.querySelector(
+			'tr[data-testid="organizations test"]'
+		);
+
+		const tableRow2 = document.querySelector(
+			'tr[data-testid="organizations test 2"]'
+		);
+
+		expect(tableRow).toHaveClass('table-active');
+
+		expect(tableRow2).toHaveClass('table-active');
 	});
 });
